requirements.txt
streamlit==1.51.0
tensorflow==2.20.0
keras==3.12.0
numpy==2.3.4
pillow==12.0.0
